LayoutTests/http/tests/xmlhttprequest/access-control-sandboxed-iframe-denied-expected.txt

CONSOLE MESSAGE: line 17: Origin null is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 17: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-sandboxed-iframe-denied.cgi due to access control checks.
This test verifies that sandboxed iframe does not have XmlHttpRequest access to its server. It will print "PASS" on success.



--------
Frame: '<!--frame1-->'
--------
PASS: Exception thrown. Sandboxed iframe XHR access was denied in 'send'. [ A network error occurred.].
